API generation script uses MDX components (#1026)
Part of #1008

This PR teaches our API generation script to generate MDX components
that will define the styling of the classes, properties, attributes,
methods, functions, and exceptions.

The script will generate custom tags to wrap all the content and store
useful information as props. This is an example of git diff for an
Attribute:
```diff
- ### dtm
- Return the system time resolution of output signals :returns: The output signal timestep in seconds. :rtype: dtm
+  ### dtm
+  <Attribute id="qiskit_ibm_provider.IBMBackend.dtm" name="dtm">
+    Return the system time resolution of output signals :returns: The output signal timestep in seconds. :rtype: dtm
+  </Attribute>
```
The script continues generating headers for every apiType that needs it,
but this will be removed in a follow-up.

### Details

The PR adds a new script named `generateMdxComponents.ts` that will
generate all the components. The main function is `processMdxComponent`
where we prepare the necessary props for each component and then return
an open tag and a closed tag ready for its use.

To process the new tags, new handlers have been added to `htmlToMd.ts`.

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>