Remove the math block's indentation from the API docs (#1123)
Turns out the math blocks (latex expressions between double dollars
`$$`) can't be indented more than once without making the pages not
render. After introducing the MDX components for classes, functions, and
attributes on #1026, we can't guarantee only one level of indentation,
and this PR adds a new function to remove the indentation as a
postprocess of all the MDX files.

Example of a math expression that was not rendering before:
![Captura desde 2024-04-03
23-35-05](https://github.com/Qiskit/documentation/assets/47946624/b3ecfb8d-b7c7-40c6-84b1-da0965ec0c15)