API generation script: fix attributes headings (#1153)
Closes #1120 

This PR changes the script to use the `name` extracted from the
`signature` of the attributes instead of using the `id` to create the
`<h3>`s. This behavior was changed in #1026, but It broke a couple of
headings.

It also fixes two attributes of `qiskit.algorithms.eigensolvers.VQD`
manually modifying the artifacts stored in
[Box](https://ibm.ent.box.com/folder/246867452622?s=tqmvw6ldivzkhh934nhufan4rg3dregr)
for versions 0.39, 0.40, 0.41, 0.42, 0.43, 0.44, 0.45 and 0.46. See more
at #1120