Upload artifacts from cron job (#1159)
We run notebooks that submit jobs twice a month to check they don't
raise errors, but we currently throw away the results. This PR uploads
the executed notebooks as an artifact so we can inspect the outputs and
update the files in this repo with more recent output.

See https://github.com/Qiskit/documentation/actions/runs/8645293171 for
an example of this working.

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>