Reorganise "Save jobs" page (#1203)
This page originally submitted a simple job using the cloud simulator as
a demonstration. Since the simulators are being deprecated, #1160
removed the job-submitting step and hardcoded the `job_id` into the
notebook. This was a bad idea because it means the notebook breaks if
not run with my IBM Quantum account.

This PR reorganises the page to get `job_id` programatically first. This
lets us avoid hardcoding the `job_id` in the notebook, which means we
can run the notebook with any IBM account that has submitted at least
one job.

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>