Use `SparsePauliOp`
The `Pauli` class does not have an `apply_layout` method, and it's not
immediately clear to me how to fix it. The notebook ran using
`SparsePauliOp` so I'm sticking with that.

Since `SparsePauliOp` doesn't have a `to_label` method, I'm sticking to
the list comprehension rather than defining them all as separate
variables.