Remove signature for the Attribute component (#1291)
This PR removes the signature prop from the Attribute component and
instead, sets `attributeTypeHint` and `attributeValue`. This change
simplifies the script that was using different props for attributes
embedded in the class page and the rest.

As a side effect, this PR also fixes an issue with some default values
not showing correctly. These values were treated as signatures and
wrongly converted into HTML by a unified plugin.

Closes #1217 #1252