Document supported windows version explicitly
In Qiskit 1.1.0 support for windows versions older than 10 was dropped
because the upstream Rust project no longer supports targeting windows
versions older than Windows 10. Previously we were not documenting the
versions of windows supported at all. But now that we have this new
constraint it's best to be explicit and say we only support windows 10
or newer.