Deduplicate `generateApiComponents` dispatch table entries (#1315)
This PR refactors the dispatch table used to prepare the props of each
API component in order to simplify the `generateApiComponents.ts`
script. The functions for "functions" and "methods" cannot be merged
trivially because they check in different ways the `isDedicatedPage`
prop. Using the same method for that gives us false positives on pages
like [this](https://docs.quantum.ibm.com/api/qiskit/0.36/execute).
Moreover, functions don't need any header in those cases.

Thanks @Eric-Arellano for the idea on #1283

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>