Use generic backend for testing (#1344)
Running notebooks with the `FakeKyoto` backend was occassionally taking
a very long time (see
https://github.com/Qiskit/documentation/pull/1307#issuecomment-2099165171).
We believe this is due to some notebooks taking advantage of the large
number of qubits it supports.

This PR switches to Qiskit's `GenericBackendV2`. This should be much
faster and gives us more fine-grained control over the test backend
properties.