Remove "warning suppression" code (#1307)
Closes #1279 
Closes #1331 

I removed the "ignore warning" code from all relevant tutorials (in
addition to VQE and CHSH) and ran them all locally on simulators to make
sure we aren't getting warnings. I also added Requirements sections to
VQE and CHSH.

Working on re-generating the output for CHSH since it currently runs on
a 27Q system, but I'm in the queue for 16 hrs.

---------

Co-authored-by: Frank Harkins <frankharkins@hotmail.co.uk>