Test that Qiskit TocGrouping is in sync with index page (#1300)
The Qiskit devs were worried about the index page from
https://github.com/Qiskit/qiskit/pull/12333 falling out of sync with our
custom grouping for `_toc.json`.

This PR adds a test to ensure that the two files are in sync. In
particular, it checks that the grouping and sequence of modules is
identical. It does not check that the label for the groups are the same,
though.

For now, we only check the dev docs because the latest docs (1.0) fail.
https://github.com/Qiskit/documentation/issues/1341 will also start
checking the latest docs once 1.1 is released.