Regeneration of API docs with modifiers (#1340)
Part of  #1287
Follow-up of #1337

This PR regenerates the API docs incorporating the changes of #1337.
This is split into two PRs to help in the review of the changes to the
script and due to the amount of files modified. Here we can find all
versions except qiskit 1.0 which was generated in #1337 to review it
more exhaustively.

This change should be merged after changing the MDX components on IQP to
support the new prop.