Host contributor license agreements (#1360)
Closes https://github.com/Qiskit/documentation/issues/796

We will host the PDFs at:

https://docs.quantum.ibm.com/open-source/qiskit-cla.pdf
https://docs.quantum.ibm.com/open-source/qiskit-corporate-cla.pdf

This folder corresponds to the new `open-source/` folder we'll want at
https://github.com/Qiskit/documentation/issues/605. We decided we don't
want to prefix the URL with `additional-resources/` because it makes the
URL too long and it makes it harder for us to reorganize docs in the
future.

We'll set up redirects for the old qiskit.org URLs.