Add check for files in `qiskit_bot.yaml` (#1375)
Closes #1330

This PR adds a check for the qiskit bot config to avoid the
`qiskit_bot.yaml` becoming stale.

The check shows two lists of files that can lead to an error:

The first one is for files that exist in the repo but don't have any
entry in the qiskit bot entry. This case will be triggered when we add
new files or we rename old ones without changing the configuration. If
we don't want to track a new file, we can add its path to the
`ALLOWED_OWNERLESS_FILES` set to avoid the error message.

The second list of files are the ones that don't exist in the repo but
have an entry in the config file. This case will be triggered by files
that change their name, removed files with the entry as a leftover, or
files where we set the config beforehand. The last ones can be added to
the `ALLOWED_INEXISTENT_FILES` set to avoid the error message.

The renamed files trigger both errors and will show the old name in the
first list, and the new name in the second one. This could be useful to
identify the old entries in the config file and to associate the files
in case we have multiple errors.

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>