Fix the properties that were incorrectly marked as methods in Qiskit 0.24-0.28 (#1387)
Closes #1352

We have 18989 misclassified properties across different versions of
Qiskit (0.24 - 0.28). To fix them, this PR added an early return to the
`getApiType()` to return the correct `apyType` in case a member has the
property modifier.

This is an example of the HTML of a property:

```html
<dl class="py method">
<dt id="qiskit.circuit.Measure.definition">
<em class="property">property </em><code class="sig-name descname">definition</code><a class="headerlink" href="#qiskit.circuit.Measure.definition" title="Permalink to this definition">¶</a></dt>
<dd><p>Return definition in terms of other basic gates.</p>
</dd></dl>
```

The script searches for the `property` modifier inside the `<dt>` tag
inside the `<dl>` tag.