External link checker: Add an HTTP header for the auth for GitHub (#1392)
Part of #823

This PR adds a new HTTP header to the external link checker to set the
auth for GitHub when we are checking GitHub links. The new header is
only used for GitHub links following what we do in the closed-source
repo.

In order to be able to get the environment variable from typescript with
`process.env.GITHUB_TOKEN`, we need to define it in the workflows. In
the case of running the script locally, the token could be undefined.

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>