Exit notebook tests early on forks (#1363)
Suggestion on how to handle notebook testing on forks. Resolves #967.

This PR proposes always failing the test on forks, with advice to merge
to another branch in this repo to test it before merging to the main
branch. While a little clunky, we very rarely get PRs from forks so this
is probably simplest for now. We can review the process if it starts
becoming a problem.

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>