Document platform support changes for Qiskit 1.1 (#1108)
In Qiskit/qiskit#12102 we're promoting the macOS arm64 environment to
tier 1 now that github actions exposes native arm64 macOS workers. This
commit updates the documentation accordingly. This should not be merged
until Qiskit 1.1.0 is released (~or 1.0.3 if we decided to backport
Qiskit/qiskit#12102~ this isn't an option because there are issues with
arm64 mac in numpy that need pretty extensive changes to workaround
which will only be in 1.1.0).

---------

Co-authored-by: Rebecca Dimock <66339736+beckykd@users.noreply.github.com>