Add script to migrate to new `guides/` structure (#1294)
This is the first step in
https://github.com/Qiskit/documentation/issues/1297.

We're going to be making last-minute changes for the docs reorg, such as
finalizing URLs at the last minute. So, it's useful to have a program
that can immediately regenerate the entire migration for us at any time.
This program will eventually generate two commits:

1. Move files to guides/
2. Update non-API internal links, like `/run/my-page` to
`/guides/my-page`

Thanks to this script being checked in to version control beforehand, we
will have less code to review in the final migration PR.

This program is not going to be permanent. We decided in
https://github.com/Qiskit/documentation/issues/1299 that we will not
generate the Workflow index pages longterm, only in this initial PR.
Instead, we'll add a validation check. So this program is only temporary
while we work on the migration.

## Follow up improvements

This PR is only the initial foundation of the migration program.
Remaining tasks:

* Finalize the page URLs
* Finalize the page organization in the left ToC and Patterns Index
pages
* Finalize the content for the new pages
* Delete the existing pages and save the change as a Git commit.
* Get CI to be green when running the migration program, such as the
link checker.
* https://github.com/Qiskit/documentation/issues/1299
* https://github.com/Qiskit/documentation/issues/1460

---------

Co-authored-by: javabster <abby.s.mitchell@gmail.com>
Co-authored-by: Arnau Casau <arnaucasau@gmail.com>