Set redirects for future deleted pages in the patterns reorg (#1497)
This PR adds a new class called `DeletedPage` that we can use to set
redirects for pages that will be removed after the reorg.

The PR also adds some redirects for the index pages of the old folders
duplicating the entries to match the two ways they can be used. An index
page can be accessed by using the name of the parent folder or by
specifying the index file. For example, the `run` index can be accessed
by going to `/run/` or `/run/index`.

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>