Have cron job create PR to update notebook outputs (#1543)
Each time the fortnightly cron job runs, I usually make a PR updating
the notebooks in this repo with the output of the cron job. This PR
automates that process. As well as saving some manual work, this also
means I can review those PRs too.

See https://github.com/frankharkins/documentation/pull/10 for an example
of this working.