Update patterns redirects file (#1594)
The PR updates the `scripts/patterns-reorg/redirects.json` file to a
more convenient format that aligns better with our middleware. Having
the `/` at the beginning of the old paths and the `/guides/` prefix in
the new paths will allow us to have a more generic redirect file that
could be used in the future for different URLs than patterns, and it
will help us to set the redirects with a simpler logic.

e.g., `url.pathname = REDIRECTS[url.pathname]`

The PR also adds a missing redirect for `/start/latest-updates`.