Add r2 to Heron entry in processor-types file (#1648)
Text from the hardware team:



    HeronR2:

This is a revision of our original Heron processor. The chip has been
redesigned to now include 156 qubits in a heavy-hexagonal lattice. While
continuing to make use of the innovations of the original Heron systems,
it also introduces a new TLS mitigation feature that controls the TLS
environment of the chip, thereby improving coherence and stability
across the whole chip.

---------

Co-authored-by: Rebecca Dimock <66339736+beckykd@users.noreply.github.com>