More QAOA notebook fixes (#1676)
We've found some more problems in the QAOA notebook after second review.
You may find it easier to review this commit-wise.

* Bit embarassed about this, but there were still two attached images in
the QAOA notebook, which are not supported by the learning platform. The
file no longer contains the string "attach" so there shouldn't be any
more.
* The `<details>` tag appeared to work, but on closer inspection it
turns out it was mangling the markdown inside the tag. I've replaced it
with a blockquote to visually separate it from the rest of the page.
* Some other minor formatting I noticed when reviewing the page.