Set up skeleton for Additional Resources dropdown (#1671)
This sets up the new URLs for the `/open-source` folder
(https://github.com/Qiskit/documentation/issues/605) and the
`/responsible-quantum-computing` page
(https://github.com/Qiskit/documentation/issues/1576).

The content is not yet synced to production. This only sets up the empty
pages and `_toc.json` so that we can develop the content more easily.

--

Note that the URLs do not have `/additional-resources/` or `/resources`
in them. We decided this in
https://github.com/Qiskit/documentation/pull/1360 so that the URLs are
shorter and more flexible to future reorganization.

`/responsible-quantum-computing` is not in a folder and does not have
`_toc.json` because it is a standalone page, similar to `/support`. If
we ever decide to expand it into a folder, we can do that by adding a
`/responsible-quantum-computing`, so long as we add an `index.mdx` file.

Migration guides will be moved over once we are ready with redirects.
(They actually technically never have to be moved.)

<img width="267" alt="Screenshot 2024-07-15 at 9 47 18 AM"
src="https://github.com/user-attachments/assets/47836660-63fa-494f-8a4b-aba3fc427059">

---------

Co-authored-by: Abby Mitchell <23662430+javabster@users.noreply.github.com>