Refactor tests to prepare for Playwright (#1745)
To unblock packaging our scripts for the Qiskit Functions repos, I plan
to migrate us to Playwright. I was having issues with getting Jest to
work with TypeScript and ES Modules (vs CommonJS modules), whereas
Playwright works out-of-the-box.

Playwright does not have `toMatchInlineSnapshot`, and `toMatch` has
different semantics; so, this switches to `toEqual`. It also removes
unnecessary `describe()` blocks.