Fix bad code in transpilation defaults (#1753)
Closes #1695 

The code snippet should actually change since the default values listed
are not what the API reference uses, and the selection of passes is
dependent on both the `optimization_level` specified as well as if a
`backend` is actually used or not.

I also think the code snippet can be cut entirely, the content is
focused on the options for the various stages and it doesn't add very
much to the explanation imo.

---------

Co-authored-by: Rebecca Dimock <66339736+beckykd@users.noreply.github.com>