Add historical redirects file (#1780)
Adds a file (and generation script) that links pages in historical API
docs to pages in the latest API docs. The banner we show on historical
docs pages can use this to link users to a more relevant page in the
latest docs, rather than always linking to the top level.

The file has the following form.

```json
{
  "package-name": {
    "version": {
      "old.page.name": "new.page.name",
    }
  }
}
```

If a page has no entry in the redirects file, we default to redirecting
to the same page in the latest API docs. E.g.
`/api/qiskit/0.33/qiskit.dagcircuit.DAGCircuit` will link to
`/api/qiskit/qiskit.dagcircuit.DAGCircuit`.

This PR only checks if a page with the same name exists in the latest
API docs and redirects to the top level if it does not. A follow up PR
could include more complicated logic to track moved classes or point to
module names when the class is removed.

<details><summary><b>Stats</b></summary>

Here are some stats to give an idea of how many pages this affects.

In total, 35% of historical API pages have a corresponding page in the
latest API versions. Looking at the breakdown per version, the majority
of recent pages will link to a more helpful page. The rest will behave
the same as before.

| Package | Page matches |
|---------|--------------|
| qiskit-ibm-provider/0.10 | 100% |
| qiskit-ibm-provider/0.7 | 100% |
| qiskit-ibm-provider/0.8 | 100% |
| qiskit-ibm-provider/0.9 | 100% |
| qiskit-ibm-runtime/0.14 | 90% |
| qiskit-ibm-runtime/0.15 | 90% |
| qiskit-ibm-runtime/0.16 | 100% |
| qiskit-ibm-runtime/0.17 | 100% |
| qiskit-ibm-runtime/0.18 | 98% |
| qiskit-ibm-runtime/0.19 | 98% |
| qiskit-ibm-runtime/0.20 | 98% |
| qiskit-ibm-runtime/0.21 | 99% |
| qiskit-ibm-runtime/0.22 | 99% |
| qiskit-ibm-runtime/0.23 | 100% |
| qiskit-ibm-runtime/0.24 | 100% |
| qiskit-ibm-runtime/dev | 100% |
| qiskit/0.19 | 23% |
| qiskit/0.24 | 23% |
| qiskit/0.25 | 20% |
| qiskit/0.26 | 20% |
| qiskit/0.27 | 20% |
| qiskit/0.28 | 21% |
| qiskit/0.29 | 21% |
| qiskit/0.30 | 21% |
| qiskit/0.31 | 21% |
| qiskit/0.32 | 21% |
| qiskit/0.33 | 33% |
| qiskit/0.35 | 34% |
| qiskit/0.36 | 34% |
| qiskit/0.37 | 34% |
| qiskit/0.38 | 34% |
| qiskit/0.39 | 34% |
| qiskit/0.40 | 33% |
| qiskit/0.41 | 33% |
| qiskit/0.42 | 33% |
| qiskit/0.43 | 39% |
| qiskit/0.44 | 53% |
| qiskit/0.45 | 56% |
| qiskit/0.46 | 57% |
| qiskit/1.0 | 97% |
| qiskit/dev | 100% |

</details>

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>