V2 additions to backend primitives section (#1794)
This PR is a proposal to partially address
https://github.com/Qiskit/documentation/issues/846. The issue was
written pre-backend primitives V2, when there were more differences in
the interface (*) between runtime and backend primitives, so proposing a
separate guide seemed reasonable at the time. Now (V2), I think the
differences are small enough to keep the backend primitives section as
an appendix to the get-started page, where I just added the usual V1-V2
tabs and a section at the end explaining differences.

This is just a rough proposal and it has been a while since I've written
a guide, so feel free to edit it to fit the style/conventions that are
used now.

---

(*) Backend primitives handled different kinds of pass managers to run
before and after parameter binding, but V2 primitives no longer do any
transpilation, so the way you run and configure backend and runtime
primitives is almost the same. I didn't add the V1 explanation for the
interface differences because V1 primitives are deprecated since qiskit
1.2 (released yesterday) and I think it would clutter the page and
confuse users to add an explanation of how they worked now.

---------

Co-authored-by: Rebecca Dimock <beckyd@us.ibm.com>