Add deprecation notice and further instructions to QuantumInstance migration guide (#1795)
This is a proposal to address
https://github.com/Qiskit/documentation/issues/1696, which doesn't
correspond with my original reply.

While updating to the changes in ISA handling doesn't seem to
complicated, I realized that updating the guide to use Primitives V2
seems very strange to me (because that is not how the migration was done
at the time). The current guide is a (in my opinion, valid) reflection
of what the motivations were with primitives V1, and the reason why it
was appreciated was because it reflected well the evolution of the code.
I also think that it is commonly understood that migration guides are
written at a specific point in time, and may require "chaining" with a
further migration guide to reach the current state of the package. So,
instead of migrating everything to V2, I propose adding a warning at the
beginning of the guide pointing to the V1 to V2 migration guide.

Once again, this is just a proposal, and I am happy to rethink this.

---------

Co-authored-by: Rebecca Dimock <beckyd@us.ibm.com>