Document new env variable in Qiskit 1.2.0 (#1836)
In Qiskit 1.2.0 a new env variable was added to control the threading
behavior of the routing and layout passes in the preset pass managers.
You can see the details of this new option in the 1.2.0 release notes or
in the PR Qiskit/qiskit#12780. This commit adds the documentation for
this feature to the local configuration guide.

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>
Co-authored-by: abbycross <across@us.ibm.com>