Fix images in Qiskit release notes for historical versions (#1848)
Before, we never copied images for the release notes file to the
historical API folder. For example, if we had the file
`public/images/qiskit/release_notes_0_0.png`, we would not copy it over
to `public/images/qiskit/1.1/release_notes_0_0.png` when we convert 1.1
from the latest release to instead be a historical release stored at
`qiskit/1.1`.

That is the right behavior for Runtime and Provider, because they only
have a single `release_notes.mdx` file. However, it does not work
correctly with Qiskit, which has one release note file per release.

This PR differentiates between the two use cases. It keeps things
generic by using our `Pkg.hasSeparateReleaseNotes` mechanism.